add log replication(AppendEntries) and leader election

+++ server/src/main/java/com/kvstore/raft/LogEntry.java
@@ -1,17 +1,17 @@
 package com.kvstore.raft;
 
 public class LogEntry {
-    private final int term;
+    private final long term;
     private final String key;
     private final String value;
 
-    public LogEntry(int term, String key, String value) {
+    public LogEntry(long term, String key, String value) {
         this.term = term;
         this.key = key;
         this.value = value;
     }
 
-    public int getTerm() {
+    public long getTerm() {
         return term;
     }
